[#188] ref: remove observing from location
