+++ config/checkstyle/suppressions.xml
@@ -95,7 +95,7 @@
     <suppress checks="MagicNumber" files="KGl33RenderFrontend.java"/>
     <suppress checks="MagicNumber" files="KGeometryUtils.java"/>
     <suppress checks="MagicNumber" files="KTextureMaker.java"/>
-    <suppress checks="MagicNumber" files="KLocation.java"/>
+    <suppress checks="MagicNumber" files="KRaycastLocationObserver.java"/>
 
 
     <suppress checks="ParameterName" files="KStbImageLwjgl.java"/>
