+++ implementations/implementations-std-level/src/test/java/io/github/darthakiranihil/konna/level/path/KAStarPathfinderPositiveTests.java
@@ -667,5 +667,7 @@ public class KAStarPathfinderPositiveTests extends KStandardTestClass {
         Assertions.assertTrue(nch.finished());
         Assertions.assertTrue(mn.finished());
 
+        Assertions.assertEquals(KVector2i.ZERO, ep.next());
+
     }
 }
